refactor: extract logic out of authform
