+++ src/lib/constants.ts
@@ -1,3 +1,7 @@
+import type { ValidAuthFormFields } from '@/types/auth'
+
 export const baseUrl = 'https://api.jikan.moe/v4'
 
 export const maxTextLengthForHiddenBlock = 285
+
+export const validationFields: ValidAuthFormFields[] = ['email', 'password']
