+++ src/pages/register.tsx
@@ -1,4 +1,4 @@
-import { AuthForm } from '@/components/AuthForm'
+import { Auth } from '@/components/Auth'
 import { PageWrapper } from '@/components/wrappers/PageWrapper'
 
 export default function Register() {
@@ -6,7 +6,7 @@ export default function Register() {
     <PageWrapper className="mt-12">
       <div className="sm:w-1/3 sm:min-w-80 sm:mx-auto">
         <h1 className="font-bold text-2xl mb-8">Create your account</h1>
-        <AuthForm />
+        <Auth />
       </div>
     </PageWrapper>
   )
