+++ src/types/auth.ts
@@ -1,0 +1,1 @@
+export type ValidAuthFormFields = 'email' | 'password'
